services dynamic pages added/ articles videos fixed/ navigation rendering dynamic
