+++ src/app/services/[documentId]/page.tsx
@@ -1,0 +1,93 @@
+import { getPage } from "@/app/lib/allDataPages";
+import Image from "next/image";
+import defaultImage from '../../../../public/luxury-office.webp';
+import { Metadata } from 'next';
+import { notFound } from 'next/navigation'
+import { StaticImport } from "next/dist/shared/lib/get-img-props";
+import { Key } from "react";
+
+// TypeScript Interfaces
+// Grouping provider metadata
+import {
+    ProviderMetadata,
+    ImageFormat,
+    MediaData,
+    TextNode,
+    LinkNode,
+    ParagraphNode,
+    ListNode,
+    ListItemNode,
+    DescriptionNode,
+    ArticleData
+} from "../../types/pageTypes";
+
+
+
+// Generates Metadata for dynamic page.
+// export async function generateMetadata({ params }: { params: { documentId: string } }): Promise<Metadata> {
+  
+//   const documentId = params.documentId;
+  
+//   const articleId = documentId;
+
+//   // Fetch the article data
+//   const json = await getArticle(articleId, true);
+//   const article = json?.data;
+
+//   if (!article) {
+//     return {
+//       title: "Article Not Found",
+//       description: "The article you're looking for doesn't exist.",
+//     };
+//   }
+
+//   return {
+//     title: article.title,
+//     description: `Read more about: ${article.title}`, // Customize this as needed
+//     openGraph: {
+//       title: article.title,
+//       description: article.description[0]?.children[0]?.text || "Private GP UK Article",
+//       url: `http://localhost:4000/${documentId}`,
+//       images: [
+//         {
+//           url: article.media?.data?.[0]?.formats?.medium?.url || 'default-image-url',
+//           width: 800,
+//           height: 600,
+//           alt: article.title,
+//         },
+//       ],
+//     },
+//   };
+// }
+
+export default async function Page({ params }: { params: { documentId: string } }) {
+  
+  const pageId = params.documentId;
+  console.log(`Here is the documentId - ${pageId} - coming in dynamically`);
+
+  if (!pageId) {
+    notFound();
+    return null;
+  }
+
+  const json: {media: any; data: ArticleData } = await getPage(pageId, true);
+  const page = json?.data;
+
+  console.log(page);
+
+  if (!page) {
+    notFound();
+    return null;
+  }
+
+    return (
+      <div className="container mx-auto px-4 mt-12 max-w-5xl">
+        <div className="md:min-h-[60vh]">
+          <div className="flex flex-col md:flex-row md:justify-between">
+            <h1 className="text-3xl mb-10 md:text-4xl font-bold">{page.Title}</h1>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
